Add Typewolf to typography section
Typewolf is a comprehensive resource containing lots of good info about web type, recommendations for free alternatives to paid fonts, pairing suggestions, and links to usage in the wild. I think this is a good addition because of its comprehensiveness. It's more of an informational resource than a dynamic tool, like some of the other pages in the typography category. Thank you!
—Michael